fix minor bugs

+++ internal/log/segment_test.go
@@ -28,7 +28,7 @@ func TestSegment(t *testing.T) {
 	for i := range 3 {
 		off, err := s.Append(want)
 		require.NoError(t, err)
-		require.Equal(t, 16+i, off)
+		require.Equal(t, uint64(16+i), off)
 		got, err := s.Read(off)
 		require.NoError(t, err)
 		require.Equal(t, want.Value, got.Value)
